+++ python/jaeger/commands/goto.py
@@ -13,7 +13,7 @@ from jaeger.utils import bytes_to_int, int_to_bytes, motor_steps_to_angle
 
 
 __ALL__ = ['InitialiseDatums', 'StartTrajectory', 'GotoAbsolutePosition',
-           'SetSpeed']
+           'SetSpeed', 'SetCurrent']
 
 
 class InitialiseDatums(Command):
@@ -88,7 +88,21 @@ class SetSpeed(Command):
     command_id = CommandID.SET_SPEED
     broadcastable = False
 
-    def __init__(self, alpha=0.0, beta=0.0, **kwargs):
+    def __init__(self, alpha=0, beta=0, **kwargs):
+
+        data = int_to_bytes(int(alpha)) + int_to_bytes(int(beta))
+        kwargs['data'] = data
+
+        super().__init__(**kwargs)
+
+
+class SetCurrent(Command):
+    """Sets the current of the alpha and beta motors."""
+
+    command_id = CommandID.SET_CURRENT
+    broadcastable = False
+
+    def __init__(self, alpha=0, beta=0, **kwargs):
 
         data = int_to_bytes(int(alpha)) + int_to_bytes(int(beta))
         kwargs['data'] = data